order

+++ CNPM-ktxUtc2Store/CNPM-ktxUtc2Store/Models/orderDetail.cs
@@ -7,7 +7,7 @@ namespace CNPM_ktxUtc2Store.Models
     {
         public int Id { get; set; }
         public int quantity { get; set; }
-        public int unitPrice { get; set; }
+        public double unitPrice { get; set; }
         public int orderId { get; set;}
         public order order { get; set;} 
         public int productId { get; set; }
